fix: crash when trying to free uninitialized pointers
In 73f55b6963dfa27a9f426c8cd34d82be89d79685, m_RSA, m_BIO, and
m_encryptionHandle were added to DestoryAuthController()
unconditionally, but they could be uninitialized when the dbus service
com.deepin.daemon.Authenticate is not present.

In CreateAuthController(), initEncryptionService() was skipped when
m_publicKey is empty, which was a result of
authControllerInter->EncryptKey failed, when the service is not present.
All these three pointers were initialized in initEncryptionService().

Log: fix a crash when trying to free uninitialized pointers.